# docs/swagger.yaml
definitions:
  ds.AddToBanknoteID:
    properties:
      banknote_id:
        type: integer
      quantity:
        type: integer
    type: object
  ds.Banknote:
    properties:
      banknote_id:
        type: integer
      currency:
        type: string
      description:
        type: string
      image_url:
        type: string
      nominal:
        type: number
      status:
        type: string
    type: object
  ds.BanknoteList:
    properties:
      banknotes_list:
        items:
          $ref: '#/definitions/ds.Banknote'
        type: array
      draft_id:
        type: integer
    type: object
...
info:
  contact: {}
paths:
  /api/banknote:
    get:
      consumes:
      - application/json
      description: Banknotes List
      parameters:
      - description: Query string to filter banknotes by nominal
        in: query
        name: banknote_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.BanknoteList'
        "500":
          description: Internal Server Error
          schema: {}
      summary: Banknotes List
      tags:
      - Banknotes
  /api/banknotes:
    delete:
      consumes:
      - application/json
      description: Deletes a banknote with the given ID
      parameters:
      - description: Banknote ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema: {}
      summary: Delete banknote by ID
      tags:
      - Banknotes
    post:
      consumes:
      - multipart/form-data
      description: Add a new banknote with image, nominal, currency
      parameters:
      - description: Banknote image
        in: formData
        name: image
        required: true
        type: file
      - description: Banknote nominal
        in: formData
        name: nominal
        required: true
        type: string
      - description: Banknote description
        in: formData
        name: description
        type: string
      - description: Banknote currency
        in: formData
        name: currency
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Add new banknote
      tags:
      - Banknotes
  /api/banknotes/:
    put:
      consumes:
      - multipart/form-data
      description: Updates a banknote with the given ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: nominal
        in: formData
        name: name
        type: string
      - description: description
        in: formData
        name: description
        type: string
      - description: currency
        in: formData
        name: IIN
        type: string
      - description: image
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema: {}
      summary: Update banknote by ID
      tags:
      - Banknotes
  /api/banknotes/{id}:
    get:
      consumes:
      - application/json
      description: Banknote By ID
      parameters:
      - description: Banknotes ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.Banknote'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Banknotes By ID
      tags:
      - Banknotes
  /api/banknotes/request:
    post:
      consumes:
      - application/json
      description: Adds a banknote to a operation request
      parameters:
      - description: Добавление банкноты
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ds.AddToBanknoteID'
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema: {}
      summary: Add banknote to request
      tags:
      - Banknotes
...
swagger: "2.0"